pyproject.toml
[tool.poetry]
name = "unc-adapter"
version = "0.1.0"
description = "Connects an isolated network environment to DeepLynx in a separate network"
authors = ["Jeren Browning <jeren.browning@inl.gov>"]

[tool.poetry.dependencies]
python = "^3.6"
deep-lynx = "*"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
